+++ pages/_app.tsx
@@ -1,14 +1,29 @@
+import { useState, useEffect } from 'react';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
-import { ThemeProvider } from 'styled-components';
 import { useRouter } from 'next/router';
+import { ThemeProvider } from 'styled-components';
+import { getCookie } from 'cookies-next';
+
+import { getUserData } from '../api/main';
 
 import GlobalStyle from '../styles/GlobalStyle';
 import Theme from '../styles/Theme';
 import Header from '../components/Header/Header';
 
 function MyApp({ Component, pageProps }: AppProps) {
+  const [userName, setUserName] = useState('');
+
   const router = useRouter();
+  const token = getCookie('token');
+
+  useEffect(() => {
+    if (token) {
+      getUserData(token).then(res => {
+        setUserName(res.data.username);
+      });
+    }
+  }, [token]);
 
   return (
     <>
@@ -18,7 +33,7 @@ function MyApp({ Component, pageProps }: AppProps) {
       </Head>
       <GlobalStyle />
       <ThemeProvider theme={Theme}>
-        {router.pathname !== '/sign-in' && <Header />}
+        {router.pathname !== '/sign-in' && <Header>{userName}</Header>}
         <Component {...pageProps} />
       </ThemeProvider>
     </>
