+++ pages/content/[contentID].tsx
@@ -1,9 +1,11 @@
 import { NextPage } from 'next';
 import React from 'react';
 
-import { getDetailData } from '../../lib/content';
+import { getDetailData } from '../../api/main';
 
 const Content: NextPage = ({ contentData }: any) => {
+  console.log(contentData);
+
   return <div>{}</div>;
 };
 
@@ -14,5 +16,5 @@ export const getServerSideProps = async (ctx: any) => {
 
   const contentData = await getDetailData(contentID);
 
-  return { props: { contentData: contentData.data.data } };
+  return { props: { contentData: contentData.data } };
 };
